Create Code for A.I. Virtual Desktop Assistant
This is the code for my A,I, virtual desktop assistant. I have also uploaded this with my .py file because if someone wants to see and if they don't have the python installed on their  system and they want to give suggestions then they can give also they can create their A.I. assistant with my codes.